Move ts to js into ts to rust
This is a step along the way so I think it should reside inside the ts
to rust folder.
