+++ src/compiler/index.ts
@@ -12,10 +12,10 @@ import {
     getStdIoType,
     logSuccess,
     printFirstBuildWarning,
-    time
+    time,
+    unwrap
 } from './utils';
 import { green } from './utils/colors';
-import { unwrap } from './utils/result';
 import {
     version as azleVersion,
     dfx_version as dfxVersion,
