+++ src/compiler/utils/index.ts
@@ -7,6 +7,7 @@ export {
 } from './global_paths';
 export { logSuccess } from './log_success';
 export { printFirstBuildWarning } from './print_first_build_warning';
+export { unwrap } from './result';
 export { time } from './time';
 
 export function getStdIoType() {
